refactor: split server object per package
