+++ pkg/svi/svi_validate.go
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
 
-// Package evpn is the main package of the application
-package evpn
+// Package svi is the main package of the application
+package svi
 
 import (
 	"go.einride.tech/aip/fieldbehavior"
