+++ pkg/vrf/vrf_validate.go
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2022-2023 Dell Inc, or its subsidiaries.
 
-// Package evpn is the main package of the application
-package evpn
+// Package vrf is the main package of the application
+package vrf
 
 import (
 	"go.einride.tech/aip/fieldbehavior"
